refactor(multiseek): wydziel logikę eksportu do multiseek_export.py
Bez zmiany zachowania — przeniesienie stałych, helperów, iteratora wierszy
i builderów odpowiedzi. Przygotowanie pod drugi wariant eksportu (FD#373).
